requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
packaging==23.2
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
watchfiles==0.21.0
websockets==12.0
